调整为整型，同时调整比较方法为按照序列比较.
git-svn-id: http://dev.bjmaxinfo.com/svn/corner/trunk/corner@3689 bda6c913-8cf9-0310-89cc-b5856a464ced

+++ src/main/java/corner/service/svn/SubversionService.java
@@ -260,7 +260,7 @@ public class SubversionService  implements IVersionService,InitializingBean{
 				
 				repository.log(new String[] {svnPath}, 
 				        0, -1, true, true,new ISVNLogEntryHandler(){
-							long seq = 1;
+							int seq = 1;
 							
 							public void handleLogEntry(SVNLogEntry logEntry) throws SVNException {
 								VersionResult result = new VersionResult();
